+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/member/service/impl/SearchMemberServiceImpl.kt
@@ -34,7 +34,7 @@ class SearchMemberServiceImpl(
                 )
             SearchMemberResponse(
                 totalPage = members.totalPages,
-                totalElements = members.totalElements.toInt(),
+                totalElements = members.totalElements,
                 data = members.content,
             )
         }
